+++ application/models/FlightInfo.php
@@ -7,9 +7,9 @@
  */
 
 // Messes up with the phpunit. Remove comments only when local.
-//function __autoload($class_name) {
-//    require_once $class_name . '.php';
-//}
+function __autoload($class_name) {
+    require_once $class_name . '.php';
+}
 
 $vars = new FleetInfo();