+++ changelog.md
@@ -77,6 +77,9 @@ Release Date: 2017-10-8
         
 ## *Version 1.0.8*
 ### Added
+        - Added tests folder. Has Bootstrap index file and PlanesTest
+        - Added Entity controller into core
+        - Made Entity model load before everything else
         
 ### Updated
         - models/FleetInfo added more information (NM)